Add moment.js library and update show entity and views
This commit adds the moment.js library to the project by adding it to the package.json file. It also updates the show entity by removing the deprecated date field and adding separate date and hour fields. The views are also updated to display the new date and hour fields.